requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
numpy>=1.24.0
gspread>=5.10.0
google-auth>=2.20.0
